Update B6.py
fix type of input

+++ B6.py
@@ -1,5 +1,5 @@
 import math
-a, b, c = map(int, input().split())
+a, b, c = map(float, input().split())
 p = (a + b + c) / 2
 s = math.sqrt(p * (p - a) * (p - b) * (p - c))
 print(s)
